requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-phonenumber-field==7.1.0
Pillow==10.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
